Moved common code out of slackEvents and check for valid token
