+++ api/config/common.go
@@ -1,9 +1,12 @@
 package common
 
 import (
+	"fmt"
 	"log"
 	"net/http"
 	"reflect"
+
+	"github.com/nlopes/slack"
 )
 
 type EnvVars struct {
@@ -17,6 +20,29 @@ type EnvVars struct {
 	NxUser                 string `split_words:"true" required:"false"`
 	NxPassword             string `split_words:"true" required:"false"`
 	GdriveFireDocFolderID  string `split_words:"true" required:"false"`
+}
+
+type SlackDAO interface {
+	SendSlackMessage(token string, attachments slack.Attachment, channel string) error
+	GetValues() []string
+}
+
+type SlackDAOImpl struct{}
+
+func (s *SlackDAOImpl) GetValues() []string {
+	return []string{"", ""}
+}
+
+func (s *SlackDAOImpl) SendSlackMessage(token string, attachments slack.Attachment, channel string) error {
+	api := slack.New(token)
+	channelID, timestamp, err := api.PostMessage(
+		channel,
+		slack.MsgOptionAttachments(attachments))
+	if err != nil {
+		return err
+	}
+	fmt.Printf("Message successfully sent to channel %s at %s", channelID, timestamp)
+	return nil
 }
 
 func SendInternalServerError(res http.ResponseWriter, err error) {
